Fix improper string handling which caused crashes
parseCloudData() made the assumption that other cloud phenomena parts
of the token would always be 3 characters or less. It was mistaken.

A real-world METAR report violated this assumption in the 01Z cycle on
February 10, 2020. It has been added to the test suite.

This change ensures that only up to 3 chars will be put into the
.other_cld_phenom buffer.